Netlify/functions/api/requirements.txt
flask
pandas
beautifulsoup4
joblib
requests
python-whois